move out configs
