+++ reloaded/src/scripts/build.mts
@@ -14,8 +14,6 @@ export const build = async () => {
     /* todo */
   );
   await builder.buildApp();
-  await $sh`mv dist/{client,worker}/assets/* dist/client/`;
-  await $sh`rmdir dist/{client,worker}/assets`;
 
   console.log("Build done!");
   console.log();
